+++ js/main.js
@@ -1,9 +1,11 @@
-import './util';
-import './variables';
-import './comment-user';
-import  './user-post';
-import './edit-img';
-import  './send-data';
-import  './upload-img';
-import  './view-images';
-import './filter-images';
+import './util.js';
+import './data.js';
+import './comment-user.js';
+import './user-post.js';
+import './edit-img.js';
+import './send-data.js';
+import './upload-img.js';
+import './miniatures.js';
+import './filter-images.js';
+import './effect-overlay.js';
+import './scale.js';
